Make the AbstractGraphTest tree more sane with only leaf nodes non-abstract and merge the tests for with and without self-loops.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=283037760
