+++ android/guava-tests/test/com/google/common/graph/ConfigurableSimpleDirectedNetworkTest.java
@@ -183,7 +183,7 @@ public class ConfigurableSimpleDirectedNetworkTest extends AbstractDirectedNetwo
       addEdge(N1, N1, E11);
       fail(ERROR_ADDED_SELF_LOOP);
     } catch (IllegalArgumentException e) {
-      assertThat(e.getMessage()).contains(ERROR_SELF_LOOP);
+      assertThat(e).hasMessageThat().contains(ERROR_SELF_LOOP);
     }
   }
