+++ guava-tests/test/com/google/common/graph/ConfigurableSimpleUndirectedNetworkTest.java
@@ -183,7 +183,7 @@ public class ConfigurableSimpleUndirectedNetworkTest extends AbstractUndirectedN
       addEdge(N1, N1, E11);
       fail(ERROR_ADDED_SELF_LOOP);
     } catch (IllegalArgumentException e) {
-      assertThat(e.getMessage()).contains(ERROR_SELF_LOOP);
+      assertThat(e).hasMessageThat().contains(ERROR_SELF_LOOP);
     }
   }
